rename mmselfsup.data to mmselfsup.structures
