+++ mmselfsup/datasets/transforms/formatting.py
@@ -7,8 +7,8 @@ from mmcv.transforms import to_tensor
 from mmcv.transforms.base import BaseTransform
 from mmengine.data import InstanceData, LabelData
 
-from mmselfsup.data import SelfSupDataSample
 from mmselfsup.registry import TRANSFORMS
+from mmselfsup.structures import SelfSupDataSample
 
 
 @TRANSFORMS.register_module()
