+++ mmselfsup/models/algorithms/cae.py
@@ -3,8 +3,8 @@ from typing import Dict, List, Optional, Union
 
 import torch
 
-from mmselfsup.data import SelfSupDataSample
 from mmselfsup.registry import MODELS
+from mmselfsup.structures import SelfSupDataSample
 from .base import BaseModel
