+++ mmselfsup/models/algorithms/densecl.py
@@ -6,8 +6,8 @@ import torch.nn as nn
 from mmengine.data import BaseDataElement
 from mmengine.model import ExponentialMovingAverage
 
-from mmselfsup.data import SelfSupDataSample
 from mmselfsup.registry import MODELS
+from mmselfsup.structures import SelfSupDataSample
 from mmselfsup.utils import (batch_shuffle_ddp, batch_unshuffle_ddp,
                              concat_all_gather)
 from .base import BaseModel
