+++ mmselfsup/models/algorithms/moco.py
@@ -5,8 +5,8 @@ import torch
 import torch.nn as nn
 from mmengine.model import ExponentialMovingAverage
 
-from mmselfsup.data import SelfSupDataSample
 from mmselfsup.registry import MODELS
+from mmselfsup.structures import SelfSupDataSample
 from mmselfsup.utils import (batch_shuffle_ddp, batch_unshuffle_ddp,
                              concat_all_gather)
 from .base import BaseModel
