+++ mmselfsup/models/algorithms/mocov3.py
@@ -4,8 +4,8 @@ from typing import Dict, List, Optional, Tuple, Union
 import torch
 import torch.nn as nn
 
-from mmselfsup.data import SelfSupDataSample
 from mmselfsup.registry import MODELS
+from mmselfsup.structures import SelfSupDataSample
 from ..utils import CosineEMA
 from .base import BaseModel
