+++ tests/test_engine/test_hooks/test_densecl_hook.py
@@ -11,10 +11,10 @@ from mmengine.model import BaseModule
 from mmengine.optim import OptimWrapper
 from torch.utils.data import Dataset
 
-from mmselfsup.data import SelfSupDataSample
 from mmselfsup.engine import DenseCLHook
 from mmselfsup.models.algorithms import BaseModel
 from mmselfsup.registry import MODELS
+from mmselfsup.structures import SelfSupDataSample
 
 
 class DummyDataset(Dataset):
