+++ tests/test_models/test_algorithms/test_simclr.py
@@ -5,8 +5,8 @@ import platform
 import pytest
 import torch
 
-from mmselfsup.data import SelfSupDataSample
 from mmselfsup.models.algorithms.simclr import SimCLR
+from mmselfsup.structures import SelfSupDataSample
 
 backbone = dict(
     type='ResNet',
