+++ tests/test_models/test_algorithms/test_simsiam.py
@@ -5,8 +5,8 @@ import platform
 import pytest
 import torch
 
-from mmselfsup.data import SelfSupDataSample
 from mmselfsup.models.algorithms import SimSiam
+from mmselfsup.structures import SelfSupDataSample
 
 backbone = dict(
     type='ResNet',
